requirements.txt
Pillow>=4.0.0,<4.0.999
requests>=2.13.0,<2.13.999

